refactor: 프레젠테이션 model 패키지에 enum을 위한 type 패키지 생성
type을 사용하는 화면 단위에서 임포트를 그에 알맞게 수정.
