+++ presentation/src/main/java/com/paykidscompose/presentation/model/type/AllowanceType.kt
@@ -1,4 +1,4 @@
-package com.paykidscompose.presentation.model
+package com.paykidscompose.presentation.model.type
 
 enum class AllowanceType{
     INCOME, // 수입
