+++ presentation/src/main/java/com/paykidscompose/presentation/model/type/QuizType.kt
@@ -1,4 +1,4 @@
-package com.paykidscompose.presentation.model
+package com.paykidscompose.presentation.model.type
 
 enum class QuizType {
     IMAGE, TEXT_CHOICE, TEXT_CHOICE_IMAGE, SHORT_ANSWER, SHORT_ANSWER_IMAGE
