+++ presentation/src/main/java/com/paykidscompose/presentation/screens/quiz/QuizScreen.kt
@@ -24,11 +24,10 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.paykidscompose.presentation.R
-import com.paykidscompose.presentation.model.QuizClearType
-import com.paykidscompose.presentation.model.QuizType
+import com.paykidscompose.presentation.model.type.QuizClearType
+import com.paykidscompose.presentation.model.type.QuizType
 import com.paykidscompose.presentation.ui.components.AppTopBar
 import com.paykidscompose.presentation.ui.components.PopupDialog
 import com.paykidscompose.presentation.ui.components.QuizResultCard
